backend_all/backend/requirements.txt
Flask
flask-cors
neo4j
gunicorn
mysql-connector-python
openai-whisper
torch
psutil
